backend/requirements.txt
fastapi
uvicorn
transformers
torch
Pillow
python-multipart
numpy
opencv-python
requests
ultralytics
